Moves message conversion functions to core.
